fix: gave mock objects separate file
